requirements.txt
# MCPVotsAGI Requirements
# =======================

# Core Dependencies
aiohttp>=3.9.0
websockets>=12.0
aiofiles>=23.2.1
aiosqlite>=0.19.0

# System & Process Management
psutil>=5.9.0
click>=8.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Database & Storage
redis>=5.0.0
ipfshttpclient>=0.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Git & Version Control
GitPython>=3.1.0

# Web & API
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# AI/ML Dependencies (optional but recommended)
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Blockchain & Trading
web3>=6.0.0
solana>=0.30.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocstrings>=0.23.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
typing-extensions>=4.8.0
pydantic>=2.0.0